+++ assets/js/synonyms/components/groups/replacements.js
@@ -46,7 +46,7 @@ export default () => {
 							// eslint-disable-next-line jsx-a11y/anchor-has-content, jsx-a11y/control-has-associated-label
 							<a
 								target="_blank"
-								href="https://www.elasticpress.io/documentation/article/how-to-disable-fuzziness/"
+								href="https://elasticprobe.com/resources/search-fuzziness-explained/"
 								rel="noreferrer"
 							/>
 						),